feat: 개개인 장소 입력 코드 리팩토링 #13
