+++ src/pages/enter-location.tsx
@@ -1,16 +1,20 @@
 import SideBar from '@/components/Sidebar';
-import { Outlet } from 'react-router-dom';
+import { Outlet, useMatch } from 'react-router-dom';
 import styled from 'styled-components';
 
 export default function EnterLocation() {
+  const locationAloneMatch = useMatch('/enter-location');
+
   return (
     <Container className="max-w-[1800px]">
       <SideBar />
       <Content>
-        <Title className="text-4xl font-medium pt-9">모두의 중간</Title>
-        <Textbox className="rounded-lg">
-          <p>상대방에게 링크를 공유하여 주소를 입력하게 하고 중간 지점을 찾아보세요!</p>
-        </Textbox>
+        <h1 className="text-4xl font-medium pt-9">모두의 중간</h1>
+        {!locationAloneMatch && (
+          <Textbox className="rounded-lg">
+            <p>상대방에게 링크를 공유하여 주소를 입력하게 하고 중간 지점을 찾아보세요!</p>
+          </Textbox>
+        )}
         <Outlet />
       </Content>
     </Container>
@@ -21,19 +25,20 @@ const Container = styled.div`
   display: flex;
   margin: 0 auto;
   width: 100vw;
+  position: relative;
+  padding-left: 100px; // Sidebar의 너비가 100px이므로 100px로 수정하였고, Sidebar의 width 변경시 동일하게 변경하면됨
+  padding-top: 10px;
 `;
 
 const Content = styled.div`
   width: 80%;
-  min-width: 1024px;
+  min-width: 1200px;
   margin: 0 auto;
   display: flex;
   flex-direction: column;
-  gap: 20px;
   align-items: center;
+  gap: 20px;
 `;
-
-const Title = styled.h1``;
 
 const Textbox = styled.div`
   width: 55%;
